+++ modules/elementary_functions/builtin/cpp/allfiniteBuiltin.cpp
@@ -1,0 +1,32 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "allfiniteBuiltin.hpp"
+#include "AllFinite.hpp"
+#include "InputOutputArgumentsCheckers.hpp"
+#include "OverloadRequired.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::ElementaryFunctionsGateway::allfiniteBuiltin(int nLhs, const ArrayOfVector& argIn)
+{
+    ArrayOfVector retval;
+    nargincheck(argIn, 1, 1);
+    nargoutcheck(nLhs, 0, 1);
+    bool needToOverload;
+    ArrayOf res = ArrayOf::logicalConstructor(AllFinite(argIn[0], needToOverload));
+    if (needToOverload) {
+        OverloadRequired("allfinite");
+    } else {
+        retval << res;
+    }
+    return retval;
+}
+//=============================================================================